remove now unused `SafariView` representable
